azure_function/requirements.txt
azure-functions==1.12.0
azure-identity==1.12.0
azure-keyvault-secrets==4.6.0
azure-storage-file-datalake==12.11.0
azure-monitor-query==1.1.1
retry==0.9.2
pandas==1.5.3
pyarrow==11.0.0
python-dotenv==1.0.0
pydantic==1.10.8
tenacity==8.2.2 